+++ tket/src/Ops/ClassicalOps.cpp
@@ -15,6 +15,7 @@
 #include "ClassicalOps.hpp"
 
 #include "OpType/OpType.hpp"
+#include "Utils/Assert.hpp"
 #include "Utils/Json.hpp"
 
 namespace tket {
@@ -85,7 +86,7 @@ static nlohmann::json classical_to_json(const Op_ptr &op, const OpType &type) {
   }
 }
 
-static std::shared_ptr<ClassicalOp> classical_from_json(
+static std::shared_ptr<ClassicalEvalOp> classical_from_json(
     const nlohmann::json &j_class, const OpType &type) {
   switch (type) {
     case OpType::MultiBit:
@@ -126,6 +127,26 @@ static std::shared_ptr<ClassicalOp> classical_from_json(
   }
 }
 
+static nlohmann::json wasm_to_json(const Op_ptr &op) {
+  nlohmann::json j_class;
+  const auto &wasm = static_cast<const WASMOp &>(*op);
+  j_class["n"] = wasm.get_n();
+  j_class["ni_vec"] = wasm.get_ni_vec();
+  j_class["no_vec"] = wasm.get_no_vec();
+  j_class["func_name"] = wasm.get_func_name();
+  j_class["wasm_uid"] = wasm.get_wasm_uid();
+  return j_class;
+}
+
+static std::shared_ptr<WASMOp> wasm_from_json(const nlohmann::json &j_class) {
+  return std::make_shared<WASMOp>(
+      j_class.at("n").get<unsigned>(),
+      j_class.at("ni_vec").get<std::vector<unsigned>>(),
+      j_class.at("no_vec").get<std::vector<unsigned>>(),
+      j_class.at("func_name").get<std::string>(),
+      j_class.at("wasm_uid").get<std::string>());
+}
+
 ClassicalOp::ClassicalOp(
     OpType type, unsigned n_i, unsigned n_io, unsigned n_o,
     const std::string &name)
@@ -138,6 +159,11 @@ ClassicalOp::ClassicalOp(
   }
 }
 
+ClassicalEvalOp::ClassicalEvalOp(
+    OpType type, unsigned n_i, unsigned n_io, unsigned n_o,
+    const std::string &name)
+    : ClassicalOp(type, n_i, n_io, n_o, name) {}
+
 nlohmann::json ClassicalOp::serialize() const {
   nlohmann::json j;
   j["type"] = get_type();
@@ -149,10 +175,31 @@ Op_ptr ClassicalOp::deserialize(const nlohmann::json &j) {
   return classical_from_json(j.at("classical"), j.at("type").get<OpType>());
 }
 
+nlohmann::json WASMOp::serialize() const {
+  nlohmann::json j;
+  j["type"] = get_type();
+  j["wasm"] = wasm_to_json(shared_from_this());
+  return j;
+}
+
+Op_ptr WASMOp::deserialize(const nlohmann::json &j) {
+  return wasm_from_json(j.at("wasm"));
+}
+
 std::string ClassicalOp::get_name(bool) const { return name_; }
 
 bool ClassicalOp::is_equal(const Op &op_other) const {
   const ClassicalOp &other = dynamic_cast<const ClassicalOp &>(op_other);
+
+  if (n_i_ != other.get_n_i()) return false;
+  if (n_io_ != other.get_n_io()) return false;
+  if (n_o_ != other.get_n_o()) return false;
+  return true;
+}
+
+bool ClassicalEvalOp::is_equal(const Op &op_other) const {
+  const ClassicalEvalOp &other =
+      dynamic_cast<const ClassicalEvalOp &>(op_other);
 
   if (n_i_ != other.get_n_i()) return false;
   if (n_io_ != other.get_n_io()) return false;
@@ -171,7 +218,8 @@ bool ClassicalOp::is_equal(const Op &op_other) const {
 
 ClassicalTransformOp::ClassicalTransformOp(
     unsigned n, const std::vector<uint32_t> &values, const std::string &name)
-    : ClassicalOp(OpType::ClassicalTransform, 0, n, 0, name), values_(values) {
+    : ClassicalEvalOp(OpType::ClassicalTransform, 0, n, 0, name),
+      values_(values) {
   if (n > 32) {
     throw std::domain_error("Too many inputs/outputs (maximum is 32)");
   }
@@ -187,6 +235,47 @@ std::vector<bool> ClassicalTransformOp::eval(const std::vector<bool> &x) const {
     y[j] = (val >> j) & 1;
   }
   return y;
+}
+
+WASMOp::WASMOp(
+    unsigned _n, std::vector<unsigned> _ni_vec, std::vector<unsigned> _no_vec,
+    const std::string &_func_name, const std::string &_wasm_uid)
+    : ClassicalOp(
+          OpType::WASM,
+          std::accumulate(
+              _ni_vec.begin(), _ni_vec.end(), decltype(_ni_vec)::value_type(0)),
+          0,
+          std::accumulate(
+              _no_vec.begin(), _no_vec.end(), decltype(_no_vec)::value_type(0)),
+          "WASM"),
+      n_(_n),
+      n_i32_(_ni_vec.size() + _no_vec.size()),
+      ni_vec_(_ni_vec),
+      no_vec_(_no_vec),
+      func_name_(_func_name),
+      wasm_uid_(_wasm_uid) {
+  unsigned sum_of_i32 =
+      std::accumulate(
+          ni_vec_.begin(), ni_vec_.end(), decltype(ni_vec_)::value_type(0)) +
+      std::accumulate(
+          no_vec_.begin(), no_vec_.end(), decltype(no_vec_)::value_type(0));
+
+  TKET_ASSERT(sum_of_i32 == n_);
+}
+
+bool WASMOp::is_equal(const Op &other) const {
+  if (other.get_type() != OpType::WASM) {
+    return false;
+  }
+
+  const WASMOp &other_wasm = dynamic_cast<const WASMOp &>(other);
+  if (other_wasm.get_n() != n_) return false;
+  if (other_wasm.get_n_i32() != n_i32_) return false;
+  if (other_wasm.get_ni_vec() != ni_vec_) return false;
+  if (other_wasm.get_no_vec() != no_vec_) return false;
+  if (other_wasm.get_func_name() != func_name_) return false;
+  if (other_wasm.get_wasm_uid() != wasm_uid_) return false;
+  return true;
 }
 
 std::string SetBitsOp::get_name(bool) const {
@@ -271,8 +360,8 @@ std::vector<bool> ExplicitModifierOp::eval(const std::vector<bool> &x) const {
   return y;
 }
 
-MultiBitOp::MultiBitOp(std::shared_ptr<const ClassicalOp> op, unsigned n)
-    : ClassicalOp(
+MultiBitOp::MultiBitOp(std::shared_ptr<const ClassicalEvalOp> op, unsigned n)
+    : ClassicalEvalOp(
           OpType::MultiBit, n * op->get_n_i(), n * op->get_n_io(),
           n * op->get_n_o(), op->get_name()),
       op_(op),